Fix electing leader from any group in cuda::memcpy_async (#6710)
We cannot rely that the passed group represents a thread block, it could be any group. So the best we can do is select the first thread. This will not lead to a uniform data path and results in worse codegen for TMA copies.

Fixes: #6601
(cherry picked from commit ace13cd23f72370b841c96ce8fca686fc6edf37f)