Implement CUDA backend for parallel `cuda::std::for_each` (#5610)
* Refactor our execution policies

We currently tag our execution policies with just an enumeration for that represents the  standard execution policies.

However, this is not sufficient for our use cases, because we also want to pass along the execution backend and the memory_direction of the algorithm.

This changes our policies so that they take an unsigned integer instead of an enumeration and then adds facilities to set and get the respective properties

* Implement parallel `cuda::std::for_each`

* Do not consider `_Backend` when comparing execution policies

* Only have once source of truth for the backend extraction